+++ tests/unit/test_utils.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 # Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You
@@ -12,7 +14,12 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
-from sagemaker.utils import get_config_value
+import pytest
+from mock import patch
+
+from sagemaker.utils import get_config_value, name_from_base, to_str, DeferredError
+
+NAME = 'base_name'
 
 
 def test_get_config_value():
@@ -32,3 +39,41 @@ def test_get_config_value():
 
     assert get_config_value('does_not.exist', config) is None
     assert get_config_value('other.key', None) is None
+
+
+def test_deferred_error():
+    de = DeferredError(ImportError("pretend the import failed"))
+    with pytest.raises(ImportError) as _:  # noqa: F841
+        de.something()
+
+
+def test_bad_import():
+    try:
+        import pandas_is_not_installed as pd
+    except ImportError as e:
+        pd = DeferredError(e)
+    assert pd is not None
+    with pytest.raises(ImportError) as _:  # noqa: F841
+        pd.DataFrame()
+
+
+@patch('sagemaker.utils.sagemaker_timestamp')
+def test_name_from_base(sagemaker_timestamp):
+    name_from_base(NAME, short=False)
+    assert sagemaker_timestamp.called_once
+
+
+@patch('sagemaker.utils.sagemaker_short_timestamp')
+def test_name_from_base_short(sagemaker_short_timestamp):
+    name_from_base(NAME, short=True)
+    assert sagemaker_short_timestamp.called_once
+
+
+def test_to_str_with_native_string():
+    value = 'some string'
+    assert to_str(value) == value
+
+
+def test_to_str_with_unicode_string():
+    value = u'åñøthér strîng'
+    assert to_str(value) == value